Move enums from common to services Naming modulename_enums.py
