+++ tests/modules/datasets/conftest.py
@@ -6,12 +6,12 @@ import pytest
 from dataall.core.environment.db.environment_models import Environment, EnvironmentGroup
 from dataall.core.organizations.db.organization_models import Organization
 from dataall.core.permissions.db.resource_policy_repositories import ResourcePolicy
-from dataall.modules.dataset_sharing.common.enums import ShareableType, PrincipalType
+from dataall.modules.dataset_sharing.services.dataset_sharing_enums import ShareableType, PrincipalType
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObject, ShareObjectItem
 from dataall.modules.dataset_sharing.services.share_permissions import SHARE_OBJECT_REQUESTER, SHARE_OBJECT_APPROVER
-from dataall.modules.datasets_base.common.enums import ConfidentialityClassification
+from dataall.modules.datasets_base.services.datasets_base_enums import ConfidentialityClassification
 from dataall.modules.datasets_base.services.permissions import DATASET_TABLE_READ
-from dataall.modules.datasets_base.db.dataset_models import Dataset, DatasetTable, DatasetStorageLocation, DatasetBucket
+from dataall.modules.datasets_base.db.dataset_models import Dataset, DatasetTable, DatasetStorageLocation
 
 
 @pytest.fixture(scope='module', autouse=True)
@@ -40,12 +40,12 @@ def dataset(client, patch_es, patch_dataset_methods):
     cache = {}
 
     def factory(
-            org: Organization,
-            env: Environment,
-            name: str,
-            owner: str,
-            group: str,
-            confidentiality: str = None
+        org: Organization,
+        env: Environment,
+        name: str,
+        owner: str,
+        group: str,
+        confidentiality: str = None
     ) -> Dataset:
         key = f'{org.organizationUri}-{env.environmentUri}-{name}-{group}'
         if cache.get(key):
@@ -268,9 +268,9 @@ def folder_fixture(db, dataset_fixture):
 @pytest.fixture(scope="module")
 def dataset_model(db):
     def factory(
-            organization: Organization,
-            environment: Environment,
-            label: str
+        organization: Organization,
+        environment: Environment,
+        label: str
     ) -> Dataset:
         with db.scoped_session() as session:
             dataset = Dataset(
