+++ tests/modules/datasets/tasks/conftest.py
@@ -2,7 +2,7 @@ import pytest
 
 from dataall.core.organizations.db.organization_models import Organization
 from dataall.core.environment.db.environment_models import Environment, EnvironmentGroup
-from dataall.modules.dataset_sharing.common.enums import ShareableType, ShareItemStatus, ShareObjectStatus, PrincipalType
+from dataall.modules.dataset_sharing.services.dataset_sharing_enums import ShareableType, ShareItemStatus, ShareObjectStatus, PrincipalType
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObjectItem, ShareObject
 from dataall.modules.datasets_base.db.dataset_models import DatasetStorageLocation, DatasetTable, Dataset, DatasetBucket
